Add meeting connect functionality
At the bare minimum, it works but theres a caveat/issue i noticed.
When two clients try to connect in the same room, they cant see each other, but the moment the server is stopped and restarted, everything works magically.
I need to find the logic that gets called when the server starts so that users can connect straightaway without any issues.